Added a hash data store implementation - it is much slower than the array store. Why? Don't know. Added really simple C code; added C stuff to gitignore. Now using int instead of int64 where possible. Fixed tiny EOF check bug and added tests for it.

+++ src/gofs/dstore/common.go
@@ -2,9 +2,9 @@ package dstore
 
 type DataStore interface {
   // Read/Write beginning at offset o from/to p
-  Read(o int64, p []byte) (int, error)
-  Write(o int64, p []byte) (int, error)
+  Read(o int, p []byte) (int, error)
+  Write(o int, p []byte) (int, error)
 
   // Returns the number of bytes stored
-  Size() int64
+  Size() int
 }
